+++ src/engine/strat_engine/tests/real.rs
@@ -33,7 +33,12 @@ impl RealTestDev {
     /// Wipe initial MiB to clear metadata.
     pub fn new(dev: Either<PathBuf, LinearDev>) -> RealTestDev {
         let test_dev = RealTestDev { dev };
-        wipe_sectors(test_dev.as_path(), Sectors(0), Bytes(IEC::Mi).sectors()).unwrap();
+        wipe_sectors(
+            test_dev.as_path(),
+            Sectors(0),
+            Bytes(u128::from(IEC::Mi)).sectors(),
+        )
+        .unwrap();
         test_dev
     }
 
@@ -44,7 +49,12 @@ impl RealTestDev {
 
     /// Teardown a real test dev
     fn teardown(self) {
-        wipe_sectors(&self.as_path(), Sectors(0), Bytes(IEC::Mi).sectors()).unwrap();
+        wipe_sectors(
+            &self.as_path(),
+            Sectors(0),
+            Bytes(u128::from(IEC::Mi)).sectors(),
+        )
+        .unwrap();
         if let Some(mut ld) = self.dev.right() {
             ld.teardown(get_dm()).unwrap();
         }
@@ -88,7 +98,7 @@ fn get_device_runs<'a>(
         }
     };
 
-    let min_size = min_size.unwrap_or_else(|| Bytes(IEC::Gi).sectors());
+    let min_size = min_size.unwrap_or_else(|| Bytes(u128::from(IEC::Gi)).sectors());
 
     assert!(max_size.is_none() || Some(min_size) <= max_size);
